mcp module refactor
