+++ backend/agent/api.py
@@ -2322,10 +2322,9 @@ async def get_custom_mcp_tools_for_agent(
             except json.JSONDecodeError:
                 logger.warning("Failed to parse X-MCP-Headers as JSON")
         
-        from mcp_module import mcp_manager
-        discovery_result = await mcp_manager.discover_custom_tools(mcp_type, mcp_config)
+        from mcp_module import mcp_service
+        discovery_result = await mcp_service.discover_custom_tools(mcp_type, mcp_config)
         
-        # Find existing MCP config for this server
         existing_mcp = None
         for mcp in custom_mcps:
             if (mcp.get('type') == mcp_type and
