+++ backend/mcp_module/__init__.py
@@ -1,14 +1,15 @@
-from .facade import MCPManager
-from .infrastructure.dependencies import get_mcp_dependencies, initialize_mcp_dependencies
-from .domain.entities import (
+from .mcp_service import (
+    MCPService,
+    mcp_service,
+
     MCPServer,
     MCPConnection, 
     MCPServerDetail,
     MCPServerListResult,
+    PopularServersResult,
     ToolExecutionResult,
     CustomMCPConnectionResult,
-)
-from .domain.exceptions import (
+
     MCPException,
     MCPConnectionError,
     MCPServerNotFoundError,
@@ -21,18 +22,14 @@ from .domain.exceptions import (
     CustomMCPError,
 )
 
-_default_dependencies = get_mcp_dependencies()
-mcp_manager = _default_dependencies.mcp_manager
-
 __all__ = [
-    "MCPManager",
-    "mcp_manager",
-    "get_mcp_dependencies",
-    "initialize_mcp_dependencies",
+    "MCPService",
+    "mcp_service",
     "MCPServer",
     "MCPConnection",
     "MCPServerDetail", 
     "MCPServerListResult",
+    "PopularServersResult",
     "ToolExecutionResult",
     "CustomMCPConnectionResult",
     "MCPException",
